Initial Restructure, Removed Google experiment specific components from classes in src/base Re-Define classes for Google Experiment in src_google Mostly functional,  parser and data copy have not yet been ammended.
