+++ airflow.cfg
@@ -642,7 +642,7 @@ smtp_ssl = True
 # Example: smtp_user = airflow
 smtp_user = primemover_py@yahoo.com
 # Example: smtp_password = airflow
-smtp_password = sghcjytieqnhqvgf
+smtp_password = ykfk rnjn lack fhtj
 smtp_port = 465
 smtp_mail_from = primemover_py@yahoo.com
 smtp_timeout = 30
