--- src/OldExperiments/2020-10-02/__init__.py
@@ -1,1 +1,0 @@
-__all__ = []
