--- src/OldExperiments/__init__.py
@@ -1,3 +1,0 @@
-from . import TestExperiment
-
-__all__ = [TestExperiment]
