+++ src/Results.py
@@ -3,12 +3,12 @@ Process results returned from primemover runner via the API
 """
 
 import src.worker.s3_parser as Parser
-from src.worker import api_wrapper
+from src.base import api_wrapper
 import json
-from datetime import datetime, timedelta
+from datetime import datetime
 from src.worker.Crawler import Crawler
 import pathlib
-import src.worker.s3_wrapper as s3
+import src.base.s3_wrapper as s3
 import zipfile
 import io
 import pandas as pd
