+++ src/__init__.py
@@ -1,2 +1,2 @@
-from . import experiment_2, worker, OldExperiments, auxiliary, BrowserLeaks, ConfigurationFunctions, experiment_2_Test, Preferences, UpdateExperiment, Results
-__all__ = [experiment_2, worker, OldExperiments, auxiliary, BrowserLeaks, ConfigurationFunctions, experiment_2_Test, Preferences, UpdateExperiment, Results]
+# from . import experiment_2, worker, OldExperiments, auxiliary, BrowserLeaks, CONFIGURATION_FUNCTIONS, experiment_2_Test, Preferences, UpdateExperiment, Results
+# __all__ = [experiment_2, worker, OldExperiments, auxiliary, BrowserLeaks, CONFIGURATION_FUNCTIONS, experiment_2_Test, Preferences, UpdateExperiment, Results]
