+++ src/worker/UpdateObject.py
@@ -1,4 +1,5 @@
-from src.worker import Agent, Proxy, ConfigureProfile, Crawler, api_wrapper
+from src.worker import Agent, Proxy, ConfigureProfile, Crawler
+from src.base import api_wrapper
 
 
 def UpdateObject(crawler_list, object_name='agent'):
@@ -6,7 +7,7 @@ def UpdateObject(crawler_list, object_name='agent'):
     object_dict = {'agent': Agent.Agent,
                    'proxy': Proxy.Proxy,
                    'crawler': Crawler.Crawler,
-                   'config': ConfigureProfile.Config}
+                   'config': ConfigureProfile.CONFIGURATION_FUNCTIONS}
     if object_name not in object_dict.keys():
         raise ValueError(f'{object_name} is not a valid object')
     if object_name == 'crawler':
@@ -42,7 +43,7 @@ def UpdateObject(crawler_list, object_name='agent'):
     elif object_name == 'config':
         for c in crawler_list:
             location = c.agent.location
-            c.configuration = ConfigureProfile.Config.from_dict(updated_dict[c.configuration.info.configuration_id], location)
+            c.configuration = ConfigureProfile.CONFIGURATION_FUNCTIONS.from_dict(updated_dict[c.configuration.info.configuration_id], location)
     else:
         raise Exception('This should not happen!')
     return crawler_list
