+++ src_google/__init__.py
@@ -1,0 +1,2 @@
+from src_google import worker
+__base__ = [worker]
